+++ clients/admin-ui/src/features/user-management/UserManagementRow.tsx
@@ -17,7 +17,11 @@ import { useRouter } from "next/router";
 import React from "react";
 import { useGetUserPermissionsQuery } from "user-management/user-management.slice";
 
+import { useAppSelector } from "~/app/hooks";
+import { selectUser } from "~/features/auth";
+import Restrict, { useHasPermission } from "~/features/common/Restrict";
 import { ROLES } from "~/features/user-management/constants";
+import { ScopeRegistryEnum } from "~/types/api";
 
 import { USER_MANAGEMENT_ROUTE } from "../../constants";
 import DeleteUserModal from "./DeleteUserModal";
@@ -30,9 +34,15 @@ interface UserManagementRowProps {
 const UserManagementRow: React.FC<UserManagementRowProps> = ({ user }) => {
   const router = useRouter();
   const deleteModal = useDisclosure();
+  const loggedInUser = useAppSelector(selectUser);
+  const isOwnProfile = loggedInUser ? loggedInUser.id === user.id : false;
+  const canEditUser =
+    useHasPermission([ScopeRegistryEnum.USER_UPDATE]) || isOwnProfile;
 
   const handleEditUser = () => {
-    router.push(`${USER_MANAGEMENT_ROUTE}/profile/${user.id}`);
+    if (canEditUser) {
+      router.push(`${USER_MANAGEMENT_ROUTE}/profile/${user.id}`);
+    }
   };
 
   const { data: userPermissions } = useGetUserPermissionsQuery(user.id ?? "", {
@@ -54,7 +64,7 @@ const UserManagementRow: React.FC<UserManagementRowProps> = ({ user }) => {
     <>
       <Tr
         key={user.id}
-        _hover={{ bg: "gray.50", cursor: "pointer" }}
+        _hover={{ bg: "gray.50", cursor: canEditUser ? "pointer" : undefined }}
         height="36px"
         data-testid={`row-${user.id}`}
       >
@@ -90,38 +100,49 @@ const UserManagementRow: React.FC<UserManagementRowProps> = ({ user }) => {
         <Td pl={0} py={1} onClick={handleEditUser}>
           {user.created_at ? new Date(user.created_at).toUTCString() : null}
         </Td>
-        <Td pr={0} py={1} textAlign="end" position="relative">
-          <ButtonGroup>
-            <Menu>
-              <MenuButton
-                as={Button}
-                size="xs"
-                bg="white"
-                data-testid="menu-btn"
-              >
-                <MoreIcon color="gray.700" w={18} h={18} />
-              </MenuButton>
-              <Portal>
-                <MenuList shadow="xl" data-testid={`menu-${user.id}`}>
-                  <MenuItem
-                    _focus={{ color: "complimentary.500", bg: "gray.100" }}
-                    onClick={handleEditUser}
-                    data-testid="edit-btn"
-                  >
-                    <Text fontSize="sm">Edit</Text>
-                  </MenuItem>
-                  <MenuItem
-                    _focus={{ color: "complimentary.500", bg: "gray.100" }}
-                    onClick={deleteModal.onOpen}
-                    data-testid="delete-btn"
-                  >
-                    <Text fontSize="sm">Delete</Text>
-                  </MenuItem>
-                </MenuList>
-              </Portal>
-            </Menu>
-          </ButtonGroup>
-        </Td>
+        <Restrict
+          scopes={[
+            ScopeRegistryEnum.USER_UPDATE,
+            ScopeRegistryEnum.USER_DELETE,
+          ]}
+        >
+          <Td pr={0} py={1} textAlign="end" position="relative">
+            <ButtonGroup>
+              <Menu>
+                <MenuButton
+                  as={Button}
+                  size="xs"
+                  bg="white"
+                  data-testid="menu-btn"
+                >
+                  <MoreIcon color="gray.700" w={18} h={18} />
+                </MenuButton>
+                <Portal>
+                  <MenuList shadow="xl" data-testid={`menu-${user.id}`}>
+                    <Restrict scopes={[ScopeRegistryEnum.USER_UPDATE]}>
+                      <MenuItem
+                        _focus={{ color: "complimentary.500", bg: "gray.100" }}
+                        onClick={handleEditUser}
+                        data-testid="edit-btn"
+                      >
+                        <Text fontSize="sm">Edit</Text>
+                      </MenuItem>
+                    </Restrict>
+                    <Restrict scopes={[ScopeRegistryEnum.USER_DELETE]}>
+                      <MenuItem
+                        _focus={{ color: "complimentary.500", bg: "gray.100" }}
+                        onClick={deleteModal.onOpen}
+                        data-testid="delete-btn"
+                      >
+                        <Text fontSize="sm">Delete</Text>
+                      </MenuItem>
+                    </Restrict>
+                  </MenuList>
+                </Portal>
+              </Menu>
+            </ButtonGroup>
+          </Td>
+        </Restrict>
       </Tr>
       <DeleteUserModal user={user} {...deleteModal} />
     </>