+++ CHANGELOG.md
@@ -40,6 +40,8 @@ The types of changes are:
 * Viewer role can now access a restricted version of the user management page [#2933](https://github.com/ethyca/fides/pull/2933)
 * Change Privacy Center email placeholder text [#2935](https://github.com/ethyca/fides/pull/2935)
 * Restricted setting Approvers as System Managers [#2891](https://github.com/ethyca/fides/pull/2891)
+* Adds confirmation modal when downgrading user to "approver" role via Admin UI [#2924](https://github.com/ethyca/fides/pull/2924)
+
 
 ### Fixed